pref: 签名服务External类接口不做文件md5校验 #3568 增加进程退出检查
